% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcross.R
\name{backcross}
\alias{backcross}
\title{backcross -------------------------------------------------------
Backcross an Indv object}
\usage{
backcross(p1, p2, c, ngen, g_map)
}
\arguments{
\item{p1}{An Indv object, the one to backcross to}

\item{p2}{An Indv object}

\item{c}{The number of chromosomes}

\item{ngen}{The number of generations of backcrossing}

\item{g_map}{The genetic map}
}
\value{
An Indv object that has been backcrossed to p1 for ngen generations
}
\description{
backcross -------------------------------------------------------
Backcross an Indv object
}
