% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcross.R
\name{backcross_pop}
\alias{backcross_pop}
\title{backcross_pop -------------------------------------------------------
 Intialize a Pop object created by backcrossing two Indv objects}
\usage{
backcross_pop(p1, pop, c, n, ngen, g_map)
}
\arguments{
\item{pop}{An Pop object}

\item{c}{The number of chromosomes}

\item{n}{The number of individuals in the population}

\item{ngen}{The number of generations of backcrossing}

\item{g_map}{The genetic map}

\item{bc_ind}{An Indv object, the one to backcross to}
}
\value{
An Pop object of individuals that have been backcrossed to p1 for ngen generations
}
\description{
backcross_pop -------------------------------------------------------
 Intialize a Pop object created by backcrossing two Indv objects
}
