% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Chrom.R
\docType{class}
\name{Chrom-class}
\alias{Chrom-class}
\alias{Chrom}
\title{Chrom}
\arguments{
\item{x}{a 'Chrom'}

\item{...}{additional 'Chrom' objects}
}
\description{
The chromosome class contains information on
 chromosome breakpoints and donors for
 for one individual chromosome.
}
\section{Slots}{

\describe{
\item{\code{chr}}{chromosome number}

\item{\code{breakpoints}}{vector of physical positions of past recombination events}

\item{\code{donors}}{vector of parental donor names for positions between breakpoints}

\item{\code{xo_no}}{integer of number of breakpoints on the chromosome}
}}

