% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Chrom.R
\docType{class}
\name{ChromPair-class}
\alias{ChromPair-class}
\alias{ChromPair}
\title{ChromPair}
\arguments{
\item{x}{a 'ChromPair'}

\item{...}{additional 'ChromPair' objects}
}
\description{
The chromosome class contains information on
chromosome breakpoints and donors for
for one individual chromosome.
}
\section{Slots}{

\describe{
\item{\code{chr}}{chromosome number}

\item{\code{h1}}{vector of physical positions of past recombination events}

\item{\code{h2}}{vector of parental donor names for positions between breakpoints}
}}

