% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_breakpoints.R
\name{count_breakpoints}
\alias{count_breakpoints}
\title{Count the number of breakpoints per individual}
\usage{
count_breakpoints(ind, hom = F)
}
\arguments{
\item{ind}{An chromosome object}

\item{hom}{Whether the indvidual is homozygous or not. Default is FALSE. If true, the number of breakpoints for only one of the chromosomes will be given}
}
\value{
The number of breakpoints in an individual. If hom is F, then it will return a vector of breakpoints on the two chromsomes, otherwise just one value will be given
}
\description{
Count the number of breakpoints per individual
}
