% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Simulate a crossover between homologous chromosomes of a parent}
\usage{
crossover(parent, recomb, c)
}
\arguments{
\item{parent}{list object in format created from chrom_init}

\item{recomb}{dataframe of genetic and physical map for chrom c}

\item{c}{The chromosome being simulated}
}
\value{
a list object of one recombinant chromosome
}
\description{
Simulate a crossover between homologous chromosomes of a parent
}
