% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Indv.R
\docType{class}
\name{Indv-class}
\alias{Indv-class}
\alias{Indv}
\alias{[,Indv-method}
\title{Indv}
\usage{
\S4method{[}{Indv}(x, i)
}
\arguments{
\item{x}{a 'Indv'}

\item{...}{additional 'Indv' objects}
}
\description{
The raw individuals class contains information on
 chromosome breakpoints and donors for one individual.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Extract Chrom by index
}}

\section{Slots}{

\describe{
\item{\code{nChr}}{number of chromosomes}

\item{\code{chroms}}{list of ChromPair objects}
}}

