% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indv_init.R
\name{indv_init}
\alias{indv_init}
\title{Intialize an Indv object}
\usage{
indv_init(chr, map, h1_donors = c("A"), h1_breakpoints = NULL,
  h2_donors = c("A"), h2_breakpoints = NULL)
}
\arguments{
\item{chr}{Number of chromosome to initialize (int); default is 10}

\item{map}{Dataframe with physical and genetic positions for chromosome trying to be simulated. Of format chr,physical_position,gentic_position}
}
\value{
A list of lists with attributes chr, h1, and h2, with h1 and h2
each containing vectors of breakpoint locations and donor names
}
\description{
Intialize an Indv object
}
