% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Indv.R
\docType{class}
\name{IndvPheno-class}
\alias{IndvPheno-class}
\alias{IndvPheno}
\alias{[,IndvPheno-method}
\title{IndvPheno}
\usage{
\S4method{[}{IndvPheno}(x, i)
}
\arguments{
\item{x}{a 'Indv'}

\item{...}{additional 'Indv' objects}
}
\description{
The raw individuals class contains information on
chromosome breakpoints and donors for one individual.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Extract Chrom by index

}}
\section{Slots}{

\describe{
\item{\code{nChr}}{number of chromosomes}

\item{\code{chroms}}{list of ChromPair objects}

\item{\code{pheno}}{Phenotype object (optional)}
}}

