% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_breaktable.R
\name{make_breaktable}
\alias{make_breaktable}
\title{Simulate a crossover between homologous chromosomes of a parent}
\usage{
make_breaktable(pop)
}
\arguments{
\item{xo}{the number of crossovers}

\item{donor1}{list object of one chromosome in format created from chrom_init, crossover starts with this one}

\item{donor2}{list object of one chromosome in format created from chrom_init}

\item{recomb}{dataframe of genetic and physical map for chrom c}

\item{c}{The chromosome being simulated}

\item{nu}{Interference intensity (optional: Default 12)}
}
\value{
a list object of one recombinant chromosome
}
\description{
Simulate a crossover between homologous chromosomes of a parent
}
