% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dh.R
\name{make_dh_pop}
\alias{make_dh_pop}
\title{make_dh_pop ------------------------------------------------------------------------
Creates a Pop object of doubled haploid (DH) lines randomly selected from input Pop. For subset of individuals,
randomly selects one of the two chromosome pairs to double}
\usage{
make_dh_pop(start_pop, n = NULL, c = 10, g_map)
}
\arguments{
\item{start_pop}{A Pop object}

\item{n}{The number of DH lines. Default is all of start_pop}

\item{c}{Chromosome number (default is 10)}

\item{g_map}{The genetic map}
}
\value{
A Pop object of n individuals that are completely homozygous
}
\description{
make_dh_pop ------------------------------------------------------------------------
Creates a Pop object of doubled haploid (DH) lines randomly selected from input Pop. For subset of individuals,
randomly selects one of the two chromosome pairs to double
}
