% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_breaktable.R
\name{make_indv_breaktable}
\alias{make_indv_breaktable}
\title{make_indv_breaktable --------------------------------------------------------
Create dataframe of breakpoint locations for a one Indv object for all chromosomes}
\usage{
make_indv_breaktable(indv, het = T)
}
\arguments{
\item{indv}{A Pop object}

\item{het}{Whether or not the chromosomes are heterozygous. Default is TRUE}
}
\value{
a dataframe with the parental donor and location of breakpoints
}
\description{
make_indv_breaktable --------------------------------------------------------
Create dataframe of breakpoint locations for a one Indv object for all chromosomes
}
