% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_magic.R
\name{make_magic}
\alias{make_magic}
\title{make_magic ------------------------------------------------------------------------
Creates a MAGIC line in funnel crossing scheme from lines in list object start_pop}
\usage{
make_magic(pop1, pop2, c = 10, g_map)
}
\arguments{
\item{pop1}{A Pop object from the starting lines of the MAGIC, should be an even number of lines
and listed in the order that crossing should occur}

\item{pop2}{A Pop object}

\item{c}{Chromosome number (default is 10)}

\item{g_map}{The genetic map}
}
\value{
An indv object that is the result of funnel crossing the lines in start_pop
}
\description{
make_magic ------------------------------------------------------------------------
Creates a MAGIC line in funnel crossing scheme from lines in list object start_pop
}
