% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_magic.R
\name{make_magic_pop}
\alias{make_magic_pop}
\title{make_magic_pop ------------------------------------------------------------------------
Creates a MAGIC population in funnel crossing scheme from lines in list object start_pop}
\usage{
make_magic_pop(start_pop, ngen = 4, c = 10, g_map, popsize)
}
\arguments{
\item{start_pop}{A Pop object from the starting lines of the MAGIC, should be an even number of lines
and listed in the order that crossing should occur}

\item{c}{Chromosome number (default is 10)}

\item{g_map}{The genetic map}

\item{popsize}{The size of the resulting population}
}
\value{
A Pop object that is the result of funnel crossing the lines in start_pop
}
\description{
make_magic_pop ------------------------------------------------------------------------
Creates a MAGIC population in funnel crossing scheme from lines in list object start_pop
}
