% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_breaktable.R
\name{make_pop_breaktable}
\alias{make_pop_breaktable}
\title{make_pop_breaktable --------------------------------------------------------
Create dataframe of breakpoint locations for a one chromsome for a Pop object}
\usage{
make_pop_breaktable(pop, n = NULL, c = 10, het = T)
}
\arguments{
\item{pop}{A Pop object}

\item{n}{The number of individuals to include in the table. Default is all of them.}

\item{c}{The chromosome number}

\item{het}{Whether or not the chromosomes are heterozygous. Default is TRUE}
}
\value{
a dataframe with the parental donor and location of breakpoints
}
\description{
make_pop_breaktable --------------------------------------------------------
Create dataframe of breakpoint locations for a one chromsome for a Pop object
}
