% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_breaktable.R
\name{make_subtable}
\alias{make_subtable}
\title{make_subtable --------------------------------------------------------}
\usage{
make_subtable(chrom, sample, het = T)
}
\arguments{
\item{chrom}{A ChromPair object}

\item{sample}{The sample name}

\item{het}{Whether or not the ChromPair is heterozygous. Default is TRUE}
}
\value{
a dataframe with the parental donor and location of breakpoints
}
\description{
Create dataframe of breakpoint locations for a single ChromPair object
}
