% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcross.R
\name{outcross}
\alias{outcross}
\title{Function to outcross population for ngen generations}
\usage{
outcross(pop, ngen, c, g_map)
}
\arguments{
\item{pop}{Starting population, pop[[i]] is the output of function offspring (list of lists)}

\item{ngen}{The number of generations of outcrossing}

\item{c}{The chromosome being simulated}

\item{g_map}{The genetic map}
}
\value{
A synthetic population of the same size as the initial population, pop that
has been randomly outcrossed for ngen generations (list of lists)
}
\description{
Function to outcross population for ngen generations
}
