% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_init.R
\name{pheno_init}
\alias{pheno_init}
\title{Intialize a Phenotype object}
\usage{
pheno_init(label = NULL, value = NULL, category = NULL)
}
\arguments{
\item{value}{phenotype value if numerical (numeric)}

\item{category}{phenotype value if categorical (character)}
}
\description{
#' @param label phenotype label (character)
}
