% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Indv.R
\docType{class}
\name{Phenotype-class}
\alias{Phenotype-class}
\alias{Phenotype}
\title{Phenotype}
\arguments{
\item{x}{a 'Phenotype'}

\item{...}{additional 'Phenotype' objects}
}
\description{
A phenotype possessed by an individual
}
\section{Slots}{

\describe{
\item{\code{label}}{phenotype label (character)}

\item{\code{value}}{phenotype value if numeric (numeric)}

\item{\code{category}}{phenotype value if categorical (character)}
}}

