% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Indv.R
\name{plot_Indv}
\alias{plot_Indv}
\title{plot_Indv ---------------------------------------------------------------
Plot all chromosomes of a single Indv object}
\usage{
plot_Indv(breaktable, het = T)
}
\arguments{
\item{breaktable}{A break table from function make_breaktable}

\item{het}{Whether or not the ChromPair is heterozygous. Default is TRUE}
}
\value{
a dataframe with the parental donor and location of breakpoints
}
\description{
plot_Indv ---------------------------------------------------------------
Plot all chromosomes of a single Indv object
}
