% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Pop.R
\docType{class}
\name{Pop-class}
\alias{Pop-class}
\alias{Pop}
\alias{[,Pop-method}
\title{Pop}
\usage{
\S4method{[}{Pop}(x, i)
}
\arguments{
\item{x}{a 'Pop'}

\item{i}{index of of individuals}

\item{...}{additional 'Pop' objects}
}
\description{
The raw individuals class contains information on
 chromosome breakpoints and donors for one individual.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Extract Indv by index
}}

\section{Slots}{

\describe{
\item{\code{nIndv}}{number of individuals}

\item{\code{poptype}}{type of population (optional)}

\item{\code{indvlist}}{list of individuals}
}}

