% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_init.R
\name{pop_init}
\alias{pop_init}
\title{Intialize a Pop object}
\usage{
pop_init(n = 4, chroms, donors = c("A", "B", "C", "D"), g_map,
  poptype = NULL)
}
\arguments{
\item{n}{Number of individuals to initialize (int)
#' @param chroms Number of chromosomes}

\item{donors}{List of names of individuals for founder population}

\item{g_map}{Dataframe with physical and genetic positions for chromosome trying to be simulated. Of format chr,physical_position,gentic_position}
}
\value{
A a Pop object of indred founder lines
}
\description{
Intialize a Pop object
}
