% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unseen.R
\name{remove_unseen}
\alias{remove_unseen}
\title{remove_unseen -------------------------------------------------------
 remove invisible crossovers from Chrom object}
\usage{
remove_unseen(chr)
}
\arguments{
\item{chr}{An Indv object}
}
\value{
A Chrom object that has had invisible crossovers removed
}
\description{
remove_unseen -------------------------------------------------------
 remove invisible crossovers from Chrom object
}
