% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self.R
\name{self}
\alias{self}
\title{self -------------------------------------------------------
 Self an Indv object}
\usage{
self(p1, c, ngen, g_map)
}
\arguments{
\item{p1}{An Indv object}

\item{c}{The number of chromosomes}

\item{ngen}{The number of generations of selfing}

\item{g_map}{The genetic map}
}
\value{
An Indv object that has been selfed for ngen generations
}
\description{
self -------------------------------------------------------
 Self an Indv object
}
