% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self.R
\name{self_pop}
\alias{self_pop}
\title{self_pop -------------------------------------------------------
 Self a Pop object}
\usage{
self_pop(pop, c, ngen, g_map)
}
\arguments{
\item{pop}{A Pop object}

\item{c}{The number of chromosomes}

\item{ngen}{The number of generations of selfing}

\item{g_map}{The genetic map}
}
\value{
An Pop object of individuals that have been selfed for ngen generations, single seed descent
}
\description{
self_pop -------------------------------------------------------
 Self a Pop object
}
