% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcross.R
\name{backcross_pop}
\alias{backcross_pop}
\title{backcross_pop -------------------------------------------------------
 Intialize a Pop object created by backcrossing two Indv objects}
\usage{
backcross_pop(p1, p2, c, n, ngen, g_map)
}
\arguments{
\item{p1}{An Indv object, the one to backcross to}

\item{p2}{An Indv object}

\item{c}{The number of chromosomes}

\item{n}{The number of individuals in the population}

\item{ngen}{The number of generations of backcrossing}

\item{g_map}{The genetic map}
}
\value{
An Pop object of individuals that have been backcrossed to p1 for ngen generations
}
\description{
backcross_pop -------------------------------------------------------
 Intialize a Pop object created by backcrossing two Indv objects
}
