% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Simulate a crossover between homologous chromosomes of a parent}
\usage{
crossover(xo, donor1, donor2, recomb, c, nu = 12)
}
\arguments{
\item{xo}{the number of crossovers}

\item{recomb}{dataframe of genetic and physical map for chrom c}

\item{c}{The chromosome being simulated}

\item{nu}{Interference intensity (optional: Default 12)}

\item{parent}{list object in format created from chrom_init}
}
\value{
a list object of one recombinant chromosome
}
\description{
Simulate a crossover between homologous chromosomes of a parent
}
