% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Simulate a crossover between homologous chromosomes of a parent}
\usage{
crossover(xo, donor1, donor2, recomb, c)
}
\arguments{
\item{xo}{the number of crossovers}

\item{donor1}{Chrom object in format created from chrom_init, crossover starts with this one}

\item{donor2}{Chrom object in format created from chrom_init}

\item{recomb}{dataframe of genetic and physical map for chrom c}

\item{c}{The chromosome being simulated}
}
\value{
a list object of one recombinant chromosome
}
\description{
Simulate a crossover between homologous chromosomes of a parent
}
