% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_breaktable.R
\name{make_breaktable}
\alias{make_breaktable}
\title{Create dataframe of breakpoint locations for simulated population}
\usage{
make_breaktable(pop, c = 10)
}
\arguments{
\item{pop}{The population of simulated lines}

\item{c}{The chromosome being simulated}
}
\value{
a dataframe with a the parental donor and location of breakpoints
}
\description{
Create dataframe of breakpoint locations for simulated population
}
