% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Pop.R
\name{plot_Pop}
\alias{plot_Pop}
\title{Create dataframe of breakpoint locations for a single ChromPair object}
\usage{
plot_Pop(breaktable, n = 10, c = 10, het = T)
}
\arguments{
\item{breaktable}{A break table from function make_breaktable}

\item{n}{The number of chromosomes to show. Default is 10}

\item{c}{The chromosome to plot}

\item{het}{Whether or not the ChromPair is heterozygous. Default is TRUE}
}
\value{
a dataframe with the parental donor and location of breakpoints
}
\description{
Create dataframe of breakpoint locations for a single ChromPair object
}
